% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applying_growth.R
\name{applyVBGF}
\alias{applyVBGF}
\title{Applies VBGF to fish length data}
\usage{
applyVBGF(Lmeas, t = 1, Lmax, Kmax, t0, t0lowbound = -0.5, silent = T)
}
\arguments{
\item{Lmeas}{a vector of fish lengths, for example, as derived from field data}

\item{t}{numeric, the number of days through which the growth trajectory of the fishes in the sample should be followed. Defaults to 1 day.}

\item{Lmax}{a vector of fish maximum species lengths, philosophically analogous to Linf}

\item{Kmax}{a vector of fish Kmax, philosophically analogous to K. Should be as outputted by the function predKmax}

\item{t0}{a vector with the theoretical size at age 0 for all fish from the data. As discussed before (multiple times, see a review in Morais and Bellwood 2018), has little biological meaning. Often constrained at 0 or using settlement size. For reef fishes it is normally larger than -0.5 (see argument \code{t0lowbound}). If not supplied, calculated using the equation from Pauly (1979).}

\item{t0lowbound}{lower bound of t0 set as default to -0.5 a (somewhat) arbitrary value that seems to make sense for reef fishes (e.g. see Choat and Robertson 2002).}

\item{silent}{would you like a function that talks to you? Here it defaults to \code{silent = TRUE}, i.e. no I would not!}
}
\value{
a vector of fish lengths represeting the size of the individuals after growing for a period of \code{t} days
}
\description{
This function applies the von Bertalanffy Growth Function to fish length data as taken from fish counts, e.g. underwater visual surveys. The von Bertalanffy Growth function describes the growth trajectory of an animal given its population growth parameters, K and Linf. For non-lethally sampled fishes, the only way to ascertain individual age is through body size. This 'operational age' represents the relative likely age of a fish, given its body size. The function first estimates the operational age from field length data and from the growth parameters (K and Linf, which, for the purpose of this package are rescaled to be Kmax and Lmax, see Morais and Bellwood (2018)), and then runs the classic VBGF to determine the length after a period of growing.
}
\references{
Depczynski, M., Fulton, C.J., Marnane, M.J., Bellwood, D.R. (2007). Life history patterns shape energy allocation among fishes on coral reefs. Oecologia 153, 111–120. doi:10.1007/s00442-007-0714-2

Morais, R.A., and Bellwood, D.R. Quantifying fish productivity, an emergent tool for managing coral reef ecosystem services. Submitted

Morais, R.A., and Bellwood, D.R. (2018). Global drivers of reef fish growth. Fish and Fisheries. 19, 874–889. doi:10.1111/faf.12297

Pauly, D. (1979). Gill size and temperature as governing factors in fish growth: A generalization of von Bertalanffy′s growth formula. Berichte aus dem Institut für Meereskunde Kiel, 63, 1–156

Choat, J. H., & Robertson, D. R. (2002). Age-based studies. In P. Sale (Ed.), Coral reef fishes: Dynamics and diversity in a complex ecosystem (pp.57–80). San Diego, CA: Academic Press. https://doi.org/10.1016/B978-012615185-5/50005-0
}
\seealso{
\code{\link{predKmax}}, \code{\link{somaGain}}
}
