% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\docType{data}
\name{db}
\alias{db}
\title{A dataset of reef fish growth trajectories}
\format{
A data frame with 1921 rows and 16 variables:
\describe{
\item{Family}{Taxonomic family}
\item{Species}{Species name}
\item{SpecCode}{Species code from FishBase}
\item{MaxSizeTL}{Maximum recorded size for the species, referring to Total Length in cm}
\item{Diet}{Dietary category, in seven levels, see original dataset}
\item{Position}{Position relative to the reef, combining horizontal and vertical components, in six levels, see original dataset}
\item{a}{Length-weight regression parameter ‘a’, estimated from the Bayesian Hierarchical Model from Froese et al. 2014 and FishBase}
\item{b}{Length-weight regression parameter ‘b’, estimated from the Bayesian Hierarchical Model from Froese et al. 2014 and FishBase}
\item{LinfTL}{Population asymptotic length in cm, converted to total length if Linf was reported in a measure other than this, see LinfType}
\item{K}{The Von Bertalanffy Growth coefficient K, as reported by the growth study}
\item{Kmax}{The standardized growth coefficient Kmax}
\item{O}{The standardized growth coefficient Ø, also termed Growth Performance Index}
\item{lon}{Longitudinal geographic coordinate of the population studied}
\item{lat}{Latitudinal geographic coordinate of the population studied}
\item{sstmean}{Mean sea surface temperature from the geographic coordinate of the population studied, obtained from Bio-ORACLE}
\item{Method}{Method used to derive the growth curve, also referred to as the ageing method, in six levels,  see original dataset}
}
}
\source{
\url{https://research.jcu.edu.au/researchdata/default/detail/a1f460789431ac7aa431212439a7ccb2/}
}
\usage{
db
}
\description{
This dataset was compiled by Morais and Bellwood (2018). The original version can be accessed in the Supporting Information - DS1 of this article including the references, locality names and complete names of the categorical traits levels. This is a streamlined version with variables formatted to be used directly within the function \code{\link{predKmax}}.
}
\references{
<Morais, R.A., and Bellwood, D.R. (2018). Global drivers of reef fish growth. Fish and Fisheries. 19, 874–889. doi:10.1111/faf.12297>

<Froese, R. (2006). Cube law, condition factor and weight-length relationships: History, meta-analysis and recommendations. Journal of Applied Ichthyology, 22, 241–253. https://doi.org/10.1111/j.1439-0426.2006.00805.x>

<Froese, R., Thorson, J. T., & Reyes, R. B. (2014). A Bayesian approach for estimating length-weight relationships in fishes. Journal of Applied Ichthyology, 30, 78–85. https://doi.org/10.1111/jai.12299>
}
\seealso{
\code{\link{predKmax}}
}
\keyword{datasets}
