% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting_growth.R
\name{predKmax}
\alias{predKmax}
\title{Predicts standardised growth parameter Kmax for reef fishes}
\usage{
predKmax(
  traits,
  dataset,
  fmod,
  params = NULL,
  niter,
  nrounds = 150,
  verbosity = 0,
  print_every_n = 1000,
  return = c("pred", "relimp", "models"),
  lowq = 0.25,
  uppq = 0.75
)
}
\arguments{
\item{traits}{data frame, intuitively, the trait object for which predictions will be made}

\item{dataset}{reference data frame from which predictions will be made. If not supplied, will use the the db dataset object from the reference above}

\item{fmod}{model formula, should be specified}

\item{params}{set of parameters to optimise the performance of xgboost. These are mapped within the function to optimise xgboost's performance. These parameters are dataset-specific so if you change the dataset, you should consider optimising these parameters using (\code{xgb.train}) within xgboost.}

\item{niter}{number of xgboost models run for the bootstrap procedure}

\item{nrounds}{maximum number of boosting interactions per model}

\item{verbosity}{should xgboost speak to you? If you do not want your workspace flooded, keep this at 0. Alternatively, check xgboost help file}

\item{print_every_n}{a friend of verbosity, treat it likewise}

\item{return}{what type of information to return. With three options: predictions (\code{pred}), relative importance of the variables used to predict (\code{relimp}), xgboost models (\code{models}). Defaults to \code{c('pred', 'relimp', 'models')}, recommended \code{c('pred')}.}

\item{lowq}{low quantile for predictions, defaults to 0.25}

\item{uppq}{upper quantile for predictions, defaults to 0.75}
}
\value{
a list including : 1) the original traits data frame with added Kmax predictions and lower and upper quantiles; 2) a table with the relative importance of the variables used in the model; and 3) a list of xgboost models
}
\description{
This function uses xgboost to predict Kmax, a standardised von Bertlanffy Growth parameter, for reef fishes. xgboost is an extremely efficient regression tree boosting system. As default, the function uses the dataset provided by Morais and Bellwood (2018), but other datasets can be supplied. Prediction is attained after executing bootstrapping xgboost estimates. Xgboost predictions are ON AVERAGE very accurate and precise, but are somewhat unstable. Because of that, I recomment any predictions to be bootstrapped at least hundreds of times (I have been using 1000 iterations, which takes something in between a few and many minutes, depending on the machine).
}
\references{
Morais, R.A., and Bellwood, D.R. (2018). Global drivers of reef fish growth. Fish and Fisheries. 19, 874–889. doi:10.1111/faf.12297
}
\seealso{
\code{\link{applyVBGF}}, \code{\link{somaGain}}
}
