% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applying_growth.R
\name{somaGain}
\alias{somaGain}
\title{Expected somatic growth in weight}
\usage{
somaGain(a, b, Lmeas, t = 1, Lmax, Kmax, t0, t0lowbound = -0.5, silent = T)
}
\arguments{
\item{a}{a vector of length-weight parameter 'a's, i.e. the intercept of the Power law}

\item{b}{a vector of length-weight parameter 'b's, i.e. the exponent of the Power law}

\item{Lmeas}{a vector of fish lengths, for example, as derived from field data}

\item{t}{numeric, the number of days through which the growth trajectory of the fishes in the sample should be followed. Defaults to 1 day.}

\item{Lmax}{a vector of fish maximum species lengths, philosophically analogous to Linf}

\item{Kmax}{a vector of fish Kmax, philosophically analogous to K. Should be as outputted by the function \code{\link{predKmax}}}

\item{t0}{a vector with the theoretical size at age 0 for all fish from the data. As discussed before (multiple times, see a review in Morais and Bellwood 2018), has little biological meaning. Often constrained at 0 or using settlement size. For reef fishes it is normally larger than -0.5 (see argument \code{t0lowbound}).}

\item{t0lowbound}{lower bound of t0 set as default to -0.5 a (somewhat) arbitrary value that seems to make sense for reef fishes.}

\item{silent}{would you like a function that talks to you? Here it defaults to \code{silent = TRUE}, i.e. no I would not!}
}
\value{
a vector of weight increments, represeting the net growth of individuals for a period of \code{t} days
}
\description{
This function calculates the expected somatic growth of a fish in weight units. It uses \code{applyVBGF} to apply the von Bertalanffy Growth Function to fish length data and then uses length-weight relationships to estimate the amount of weight that this somatic growth generated. It is the basis for productivity estimates.
}
\seealso{
\code{\link{applyVBGF}}, \code{\link{predKmax}}
}
