% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary.R
\name{tidytrait}
\alias{tidytrait}
\title{Apply standardised factors to growth prediction data frame}
\usage{
tidytrait(traits, dataset, levels.dataset = TRUE, include.method = FALSE)
}
\arguments{
\item{traits}{data frame, intuitively the trait object for which predictions will be made}

\item{dataset}{reference data frame, intuitively the db dataset object from which predictions will be made}

\item{levels.dataset}{Includes variable levels required later on for predicting growth. Defaults to TRUE}

\item{include.method}{Should include a variable for the method used to derive the growth curves? Because db in Morais and Bellwood (2018) includes growth curves obtained from different methods, this argument makes sure that predictions are made based on the curves obtained by ageing from otoliths only. Defaults to \code{FALSE}, assuming that the variable is already present in traits}
}
\value{
a tidied up traits data frame with factors ready to be used to predict growth
}
\description{
Makes sure that the trait object (species for which to predict growth) has factors perfectly matching the db object (species from which to predict growth)
}
\examples{
data (db)
levels (db$Diet)

}
\references{
<Morais, R.A., and Bellwood, D.R. (2018). Global drivers of reef fish growth. Fish and Fisheries. 19, 874–889. doi:10.1111/faf.12297>
}
