% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predicting_mortality.R
\name{predM}
\alias{predM}
\title{Predicts natural mortality rates M for reef fishes}
\usage{
predM(Lmeas, t = 1, Lmax, Kmax, Lr, temp, p = 0.5, method = c("Pauly"))
}
\arguments{
\item{Lmeas}{a vector of fish sizes, for example, as derived from field data}

\item{t}{numeric, the number of days through which the growth trajectory of the fishes in the sample should be followed. Defaults to 1 day. In most of the fisheries literature, M is scaled by year instead (i.e. use 365 in that case)}

\item{Lmax}{a vector of fish maximum species lengths, philosophically analogous to Linf}

\item{Kmax}{a vector of fish Kmax, philosophically analogous to K. Should be as outputted by the function \code{\link{predKmax}}}

\item{Lr}{a vector of sizes at settlement. Only relevant for \code{method = 'Function'}}

\item{temp}{a vector of environmental temperatures to which the fish are exposed. Only relevant for \code{method = c('Pauly', 'Function')}}

\item{p}{proportion of the body size range of a species to scale the function to be applied to M at the species-population level. Defaults to 0.5, meaning that M at the individual level is equal to M at the species/population level for fishes in the mid point of their body size range}

\item{method}{method to be used for predicting mortality rates. Can be either \code{'Pauly'}, \code{'Gislason'} or \code{'Function'}. Pauly's equation predicts M at the species/population level from Linf (Lmax here), K (Kmax here) and temperature. Gislason (et al.)'s equation predicts M at the individual-level by incorporating individual size (Lmeas), Lmax and Kmax. The 'Function' method (see Morais and Bellwood, submitted) includes a functional relationship with individual body size to the population-level estimate of M from Pauly. Because there is currently no recommendation on which method should be used for estimating size-specific mortality rates, \code{'Pauly'} is the default, although it should (rather paradoxically) potentially NOT be used.}
}
\description{
This function uses empirical relationships to predict the likely mortality rates, M, of a fish population or individual (depending on the method, see below). M is the instantaneous mortality rate, the exponent describing the decreasing number of individuals from a cohort through time and, by extension, also describing the cumulative per capita mortality rate of an individual.
}
\references{
Morais, R.A., and Bellwood, D.R. Quantifying fish productivity, an emergent tool for managing coral reef ecosystem services. Submitted

Gislason H, Daan N, Rice JC, and Pope JG. 2010. Size, growth, temperature and the natural mortality of marine fish. Fish Fish 11: 149–58

Pauly, D. (1980). On the interrelationships between natural mortality, growth parameters, and mean environmental temperature in 175 fish stocks. ICES J. Mar. Sci. 39, 175–192
}
\seealso{
\code{\link{somaLoss}}, \code{\link{applyMstoch}}
}
