% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfishprod-package.R
\docType{package}
\name{rfishprod-package}
\alias{rfishprod}
\alias{rfishprod-package}
\title{rfishprod: Estimating Reef Fish Productivity}
\description{
Implements a framework to generate individual-level estimates
    of fish productivity, with a focus on reef fishes. This individual
    approach works by combining estimates of somatic growth through the
    von Bertalanffy Growth Model, and deterministic or stochastic natural
    mortality using instantaneous mortality rates.
}
\section{\code{rfishprod} functions}{

Use the functions \code{predKmax} and \code{predM} to predict growth and mortality trajectories, respectively. Then, use \code{somaGain}, \code{applyMstoch} and \code{somaLoss} to position your fishes in these trajectories and estimate components of productivity!
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/renatoamorais/rfishprod}
}

}
\author{
\strong{Maintainer}: Renato Morais \email{renatomoraisaraujo@gmail.com} (\href{https://orcid.org/0000-0003-4652-6676}{ORCID})

}
\keyword{internal}
